Fix vertical centering of footer text
It was geometrically centered.. but it did not look right. This must have someting
to do with how Helvetica shoots above and below. Because there is almost no letters
shooting down, the text appeared to be too far up.

This is now compensated with different padding above and below the copyright
text.