Add static query to header and footer
When I first tried it, I could not get it to work. I guess it was because of a
syntax issue. I was using useStaticQuery(graphql()). This leaves the graphql
call in the code and fails.

The correct way to use it is useStaticQuery(graphql`...`)
For some reason this is the syntax I had used when I moved the calls to the
layout component.