Move external spacing from contact component to index page
The external spacing should be determined on the page basis. The component can not really know what
it needs.

Also the definition as a section has been moved from the component to the page. Pretty much for the same
reason that the spacing is applied from the parent. The meaning of the components depends on the context.
This is a bit different than a button or something where the semanitic meaning is always the same.
;