ButtonLinks are local when starting with `/` or `#`
I am actually not using the Button link to link on the site at all so far.
But I also made a mistake when implementing. I though that checking for `http` at the
start would be enough for external links. This already failed when creating an email link.

Now the logic is inverted and by default external is assumed and only when starting with
`/` or `#` (for heading links) will the local link be used.